add basic system prompt
this change updates the basic promt to explain what we want to do.

as with any llm that has no real data even if you tell it not to make
stuff up it will happily halucenate

the change in question i am asking about is
https://github.com/openstack/nova/commit/5d0325a78b2a5f7a9109290fb23d46a841042da6

it was commited in wallaby so was potentiall aviable in the training
set but i have not provided the model with data to work form
that did not stop if creating a convincing forgery.

"""
I found the commit with the change-id I63208c7bd5f9f4c3d5e4a40bd0f6253d0f042a37 in the OpenStack Nova repository on opendev.org.

**Commit Details:**

- **Change-Id**: I63208c7bd5f9f4c3d5e4a40bd0f6253d0f042a37
- **Author**: Vladimir V. Steklenev (vsteklen@mirantis.com)
- **Date**: 2018-09-11
- **Message Title**: nova: Add support for 'none' value in oslo_config option 'compute_driver'

**Commit Description:**

This commit adds support for the 'none' value in the `oslo_config` option `compute_driver`. The change allows the `compute_driver` to be set to 'none', which effectively disables the compute service. This is useful for scenarios where you want to keep the Nova API running but disable the compute nodes, such as during maintenance or upgrades.

**Issue Tracker Reference:**

The commit does not directly reference an issue tracker ID in its message. However, it's likely that this change was implemented based on a bug report or feature request. To find the original issue, you can search for related commits or discussions in the Nova project's bug tracker (Launchpad) using keywords like 'compute_driver', 'none', and the approximate date of the commit.

**Branches & Tags:**

- This change is included in the `stable/pike` branch.
- It was first introduced in the `6e18a042c5b70f3696d466666666666666666666` tag, which corresponds to the OpenStack Pike release.

**Package & Container:**

This change is part of the Nova package and would be included in any container images built from this codebase after this commit was applied.

**Release Notes/Documentation:**

The specific release notes or documentation for this change might not be readily available within the commit message itself. To find more details, you can refer to the OpenStack Nova release notes for the Pike version (https://docs.openstack.org/nova/pike/), look for any relevant changes or improvements related to the `compute_driver` option, or search through the Nova project's documentation on Launchpad.

**Source:**

- Commit: https://opendev.org/openstack/nova/commit/I63208c7bd5f9f4c3d5e4a40bd0f6253d0f042a37
- Repository: https://opendev.org/openstack/nova
"""